+++ src/Symfony/Component/Security/Core/Authorization/Voter/VoteInterface.php
@@ -24,7 +24,7 @@ interface VoteInterface
 
     public function isGranted(): bool;
 
-    public function isAbstain(): bool;
+    public function isAbstainer(): bool;
 
     public function isDenied(): bool;