+++ src/Symfony/Component/Security/Core/CHANGELOG.md
@@ -9,6 +9,7 @@ CHANGELOG
  * Add `OfflineTokenInterface` to mark tokens that do not represent the currently logged-in user
  * Deprecate `UserInterface::eraseCredentials()` and `TokenInterface::eraseCredentials()`,
    erase credentials e.g. using `__serialize()` instead
+ * Add the ability for voter to return decision reason and a score by passing a `Vote` object
 
 7.2
 ---
@@ -17,7 +18,6 @@ CHANGELOG
  * Add `$token` argument to `UserCheckerInterface::checkPostAuth()`
  * Deprecate argument `$secret` of `RememberMeToken`
  * Deprecate returning an empty string in `UserInterface::getUserIdentifier()`
- * Add the ability for voter to return decision reason and a score by passing a `Vote` object
 
 7.0
 ---